feat(example): add macOS camera/mic to browser streaming example
Add examples/macos_camera/ — real-time macOS camera and microphone
streaming to browser via nanortc WebRTC with multi-viewer support.

- AVFoundation capture (av_capture.h/m): camera + mic with auto preset
  selection (480p/720p/1080p based on config)
- VideoToolbox H.264 encoder (vt_encoder.h/m): hardware encoding with
  AVCC-to-Annex-B conversion, real-time mode, forced keyframe support
- Multi-session architecture: up to 4 simultaneous browser viewers,
  shared capture source with per-viewer nanortc_t + UDP socket
- Signaling server host mode: persistent host (pid=0) with N viewers,
  backward-compatible legacy 2-peer mode preserved
- HTTP signaling extensions: http_sig_join_as_host(), http_sig_send_to(),
  http_sig_recv_from() for host-mode routing
- Browser UI: unified video+audio controls, fullscreen, default muted,
  proper ontrack with e.streams[0] fallback
- nanortc fixes: out_head/out_tail uint8_t→uint16_t for queue >128,
  app-level NANORTC_CONFIG_FILE override for HD video queue sizing

+++ include/nanortc.h
@@ -511,8 +511,8 @@ struct nanortc {
 
     /* Output queue (simple ring buffer) */
     nanortc_output_t out_queue[NANORTC_OUT_QUEUE_SIZE];
-    uint8_t out_head;
-    uint8_t out_tail;
+    uint16_t out_head;
+    uint16_t out_tail;
 
     /* Scratch buffer for STUN encode/decode.
      * Sans I/O contract: caller must drain outputs before next handle_receive. */
